factor out questions query to a model
